requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

httpx
azure-functions
pandas
python-dotenv
BeautifulSoup4
openpyxl
xlsxwriter
supabase
psycopg2-binary
tqdm
langchain
openai
langgraph
pytest
